adding module for searching ncdps id's

+++ resources/views/partials/clients/client-index.blade.php
@@ -58,7 +58,9 @@
                                                 
                                                 <a href="/client/{{ $client->id }}/edit" class="btn btn-primary btn-sm">Edit</a>
                                                 {{--  <button class="btn btn-primary"> View Notes</button>  --}}
+                                               @can('eat', $client) 
                                                 <a href="/delete-client/{{ $client->id }}" id="delete" class="btn btn-danger btn-sm"> <span class="glyphicon glyphicon-remove"><strong> X </strong></span></a>
+                                                @endcan
                                         </td>
                                     </tr>
                                 @endforeach
